requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_google_genai
youtube_transcript_api
pathlib
crewai
crewai_tools
load_dotenv
langchain-huggingface